add safety check, use bigger data types

+++ ComplexityAlgorithms.hpp
@@ -29,9 +29,9 @@ class ComplexityAlgorithms
 public:
     bool randomWalk(satFormula *S, int t);
     vector<unsigned short> solveKP(knapsackProblem*);
-    unsigned short solveKP2(knapsackProblem *);
+    int solveKP2(knapsackProblem *);
     int min(int a, int b){ return a < b ? a : b; }
-    unsigned short max(unsigned short a, unsigned short b){ return a > b ? a : b; }
+    int max(int a, int b){ return a > b ? a : b; }
 
 private:
     bool satisfied(struct satFormula *) const ;
